Added scripts and loaders for links page
Earlier, all the assets and information was
present locally and was used for developing the
frontend. Now, scripts have been added for
fetching of user data from the server, using the
username or the session token provided, and if the
user or session token is invalid or not found,
the site redirects to the login page.