Added Login Form page and integrated it with the routing
Earlier there was only the public links page,
which is now preceded by the login page where the
user is supposed to enter their credentials and
get authenticated before they can view their links
Eventually this page will end up taking the user
to the edit page and not the links page, as the
edit page is only going to be available to the
user themselves whereas the links page will be
available publicly.

The scripts needed for the login page, such as
finding the user, checking their credentials and
requesting session tokens for remember me
functionality have been added as well.