Added script for local storage, integrated session tokens api for logins
Earlier, logins were performed exclusively using
username and password, and the remember me
functionality did not work. Now, session tokens
are stored in the local storage and can be used to
automatically login, using the session token api,
and the local storage api. The tokens are
periodically cleaned from the db as well, which
makes sure all tokens are up to date.