Added vercel.json as an attempt to fix failed dynamic routing
The user links page uses dynamic routes to display
the links associated with any username, which
works locally but fails on vercel. This change has
been made to fix this issue, where a vercel.json
file has been added to the root directory to add
custome rewrites to the app. This was suggested by
this article on the vercel website:
https://vercel.com/docs/frameworks/vite